Split home page into couple components
